Rename the file of filters
